Refactoring. add login from Id

+++ database/migrations/2017_09_11_123921_add_fb_dield.php
@@ -15,7 +15,7 @@ class AddFbDield extends Migration
     {
         if(Schema::hasTable('users')) {
             Schema::table('users', function (Blueprint $table) {
-                $table->integer('fb_id');
+                $table->bigInteger('fb_id');
             });
         }
     }
